Finished the twitter like app

+++ src/endpoints/messages/delete.js
@@ -1,11 +1,13 @@
 module.exports = {
     method:'post',
-    endpoint:'/messages/delete/:username/:content',
+    endpoint:'/messages/delete/:content',
     cmd: async (db, req, res) => {
         const content = req.params.content;
-        const username = req.params.username;
+        const username = req.cookies.user;
 
-        console.log(username, content);
+        if(!username)
+            return res.sendStatus(404);
+
 
         const deleteMsg = await db.msgs.destroy({where:{username: username, content: content }});
