+++ src/endpoints/user/delete.js
@@ -1,29 +1,33 @@
 module.exports =  {
     method: 'post',
-    endpoint: '/user/delete/:username/:password',
+    endpoint: '/user/delete',
     cmd: async (db, req, res) => {
-        const username = req.params.username;
-        const pass = req.params.password;
+        const username = req.cookies.user || null;
+
+        if(!username)
+            return res.sendStatus(404);
 
         const deleteUser = await db.users.destroy({
-            where: {
-                username: username,
-                password: pass
-            }
+            where: { username: username }
         });
 
-        if(!deleteUser) {
+        if(!deleteUser)
             return res.sendStatus(404);
-        }
 
-        const deleteMsg = await db.msgs.destroy({
+        await db.msgs.destroy({
             where: {
                 username: username
             }
         });
 
-        if(!deleteMsg)
-            return res.sendStatus(404);
+        const currentUser = req.cookies.user || null;
+
+        if(currentUser)
+            res.clearCookie("user", {
+                httpOnly: true,
+                sameSite: true
+            })
+
 
         return res.sendStatus(200);
     }
